W172d: btlFormation FOUR MATCHes -- the entire 002CE/002CF family closed. 002CEB20 nd42 -> 0, 002CEC30 nd12 -> 0, 002CEF10 nd18 -> 0, 002CF1D0 nd16 -> 0. All four had been cleared from over-window in the previous pass by replacing invented array zero-fills, u64 masks/widths and untyped pointer/call paths with sibling-grounded typed selection logic; this pass finished them. Also 002BAD60 nd17 -> 9.
File 466 -> 470 MATCH (90.4%). Over-window unchanged at 19 with no additions, LF-only endings preserved (0 CRLF, 15,140 LF), 46/46 pragma brackets and the retained opt_common_subs untouched.

Clear demonstration that clearing a window and closing the function are two stages of one job: identify and delete the invented work first, then the residual becomes tractable.

+++ src/Event/Fcl/fclCombineMisc.c
@@ -2114,13 +2114,13 @@ u32 FUN_003d7ac0(u8 param_1,u8 param_2,u16 *param_3)
 
   int iVar6;
 
-  u32 uVar7;
+  u16 uVar7;
 
-  u32 uVar8;
+  u16 uVar8;
 
-  u32 uVar9;
+  u16 uVar9;
 
-  u32 uVar10;
+  u16 uVar10;
 
   int iVar11;
 
@@ -2303,13 +2303,13 @@ u32 FUN_003d7da0(u8 param_1,u8 param_2,u16 *param_3)
 
   int iVar6;
 
-  u32 uVar7;
+  u16 uVar7;
 
-  u32 uVar8;
+  u16 uVar8;
 
-  u32 uVar9;
+  u16 uVar9;
 
-  u32 uVar10;
+  u16 uVar10;
 
   int iVar11;
